[feat][1.0.0]进房协议完成

+++ poker-common/src/main/java/org/cloud/entity/poker/BasePacket.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class BasePacket {
-    private short protoCode;
+    private byte protoCode;
     private byte[] bytes;
 
 }
